@@ -40,7 +40,7 @@ void main() {
   });
 
   testWidgets('Initial url', (WidgetTester tester) async {
-    WebViewController controller;
+    late WebViewController controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://youtube.com',
@@ -60,7 +60,7 @@ void main() {
     ));
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformWebView.javascriptMode, JavascriptMode.unrestricted);
 
@@ -72,7 +72,7 @@ void main() {
   });
 
   testWidgets('Load url', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -83,13 +83,13 @@ void main() {
 
     expect(controller, isNotNull);
 
-    await controller.loadUrl('https://flutter.io');
+    await controller!.loadUrl('https://flutter.io');
 
-    expect(await controller.currentUrl(), 'https://flutter.io');
+    expect(await controller!.currentUrl(), 'https://flutter.io');
   });
 
   testWidgets('Invalid urls', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -100,19 +100,18 @@ void main() {
 
     expect(controller, isNotNull);
 
-    expect(() => controller.loadUrl(null), throwsA(anything));
-    expect(await controller.currentUrl(), isNull);
+    expect(await controller!.currentUrl(), isNull);
 
-    expect(() => controller.loadUrl(''), throwsA(anything));
-    expect(await controller.currentUrl(), isNull);
+    expect(() => controller!.loadUrl(''), throwsA(anything));
+    expect(await controller!.currentUrl(), isNull);
 
     // Missing schema.
-    expect(() => controller.loadUrl('flutter.io'), throwsA(anything));
-    expect(await controller.currentUrl(), isNull);
+    expect(() => controller!.loadUrl('flutter.io'), throwsA(anything));
+    expect(await controller!.currentUrl(), isNull);
   });
 
   testWidgets('Headers in loadUrl', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -126,13 +125,13 @@ void main() {
     final Map<String, String> headers = <String, String>{
       'CACHE-CONTROL': 'ABC'
     };
-    await controller.loadUrl('https://flutter.io', headers: headers);
-    expect(await controller.currentUrl(), equals('https://flutter.io'));
+    await controller!.loadUrl('https://flutter.io', headers: headers);
+    expect(await controller!.currentUrl(), equals('https://flutter.io'));
   });
 
   testWidgets("Can't go back before loading a page",
       (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -143,13 +142,13 @@ void main() {
 
     expect(controller, isNotNull);
 
-    final bool canGoBackNoPageLoaded = await controller.canGoBack();
+    final bool canGoBackNoPageLoaded = await controller!.canGoBack();
 
     expect(canGoBackNoPageLoaded, false);
   });
 
   testWidgets("Clear Cache", (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -159,15 +158,15 @@ void main() {
     );
 
     expect(controller, isNotNull);
-    expect(fakePlatformViewsController.lastCreatedView.hasCache, true);
+    expect(fakePlatformViewsController.lastCreatedView!.hasCache, true);
 
-    await controller.clearCache();
+    await controller!.clearCache();
 
-    expect(fakePlatformViewsController.lastCreatedView.hasCache, false);
+    expect(fakePlatformViewsController.lastCreatedView!.hasCache, false);
   });
 
   testWidgets("Can't go back with no history", (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -178,13 +177,13 @@ void main() {
     );
 
     expect(controller, isNotNull);
-    final bool canGoBackFirstPageLoaded = await controller.canGoBack();
+    final bool canGoBackFirstPageLoaded = await controller!.canGoBack();
 
     expect(canGoBackFirstPageLoaded, false);
   });
 
   testWidgets('Can go back', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -196,15 +195,15 @@ void main() {
 
     expect(controller, isNotNull);
 
-    await controller.loadUrl('https://www.google.com');
-    final bool canGoBackSecondPageLoaded = await controller.canGoBack();
+    await controller!.loadUrl('https://www.google.com');
+    final bool canGoBackSecondPageLoaded = await controller!.canGoBack();
 
     expect(canGoBackSecondPageLoaded, true);
   });
 
   testWidgets("Can't go forward before loading a page",
       (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -215,13 +214,13 @@ void main() {
 
     expect(controller, isNotNull);
 
-    final bool canGoForwardNoPageLoaded = await controller.canGoForward();
+    final bool canGoForwardNoPageLoaded = await controller!.canGoForward();
 
     expect(canGoForwardNoPageLoaded, false);
   });
 
   testWidgets("Can't go forward with no history", (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -232,13 +231,13 @@ void main() {
     );
 
     expect(controller, isNotNull);
-    final bool canGoForwardFirstPageLoaded = await controller.canGoForward();
+    final bool canGoForwardFirstPageLoaded = await controller!.canGoForward();
 
     expect(canGoForwardFirstPageLoaded, false);
   });
 
   testWidgets('Can go forward', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -250,15 +249,15 @@ void main() {
 
     expect(controller, isNotNull);
 
-    await controller.loadUrl('https://youtube.com');
-    await controller.goBack();
-    final bool canGoForwardFirstPageBacked = await controller.canGoForward();
+    await controller!.loadUrl('https://youtube.com');
+    await controller!.goBack();
+    final bool canGoForwardFirstPageBacked = await controller!.canGoForward();
 
     expect(canGoForwardFirstPageBacked, true);
   });
 
   testWidgets('Go back', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://youtube.com',
@@ -270,19 +269,19 @@ void main() {
 
     expect(controller, isNotNull);
 
-    expect(await controller.currentUrl(), 'https://youtube.com');
+    expect(await controller!.currentUrl(), 'https://youtube.com');
 
-    await controller.loadUrl('https://flutter.io');
+    await controller!.loadUrl('https://flutter.io');
 
-    expect(await controller.currentUrl(), 'https://flutter.io');
+    expect(await controller!.currentUrl(), 'https://flutter.io');
 
-    await controller.goBack();
+    await controller!.goBack();
 
-    expect(await controller.currentUrl(), 'https://youtube.com');
+    expect(await controller!.currentUrl(), 'https://youtube.com');
   });
 
   testWidgets('Go forward', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://youtube.com',
@@ -294,23 +293,23 @@ void main() {
 
     expect(controller, isNotNull);
 
-    expect(await controller.currentUrl(), 'https://youtube.com');
+    expect(await controller!.currentUrl(), 'https://youtube.com');
 
-    await controller.loadUrl('https://flutter.io');
+    await controller!.loadUrl('https://flutter.io');
 
-    expect(await controller.currentUrl(), 'https://flutter.io');
+    expect(await controller!.currentUrl(), 'https://flutter.io');
 
-    await controller.goBack();
+    await controller!.goBack();
 
-    expect(await controller.currentUrl(), 'https://youtube.com');
+    expect(await controller!.currentUrl(), 'https://youtube.com');
 
-    await controller.goForward();
+    await controller!.goForward();
 
-    expect(await controller.currentUrl(), 'https://flutter.io');
+    expect(await controller!.currentUrl(), 'https://flutter.io');
   });
 
   testWidgets('Current URL', (WidgetTester tester) async {
-    WebViewController controller;
+    WebViewController? controller;
     await tester.pumpWidget(
       WebView(
         onWebViewCreated: (WebViewController webViewController) {
@@ -322,20 +321,20 @@ void main() {
     expect(controller, isNotNull);
 
     // Test a WebView without an explicitly set first URL.
-    expect(await controller.currentUrl(), isNull);
+    expect(await controller!.currentUrl(), isNull);
 
-    await controller.loadUrl('https://youtube.com');
-    expect(await controller.currentUrl(), 'https://youtube.com');
+    await controller!.loadUrl('https://youtube.com');
+    expect(await controller!.currentUrl(), 'https://youtube.com');
 
-    await controller.loadUrl('https://flutter.io');
-    expect(await controller.currentUrl(), 'https://flutter.io');
+    await controller!.loadUrl('https://flutter.io');
+    expect(await controller!.currentUrl(), 'https://flutter.io');
 
-    await controller.goBack();
-    expect(await controller.currentUrl(), 'https://youtube.com');
+    await controller!.goBack();
+    expect(await controller!.currentUrl(), 'https://youtube.com');
   });
 
   testWidgets('Reload url', (WidgetTester tester) async {
-    WebViewController controller;
+    late WebViewController controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -346,7 +345,7 @@ void main() {
     );
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformWebView.currentUrl, 'https://flutter.io');
     expect(platformWebView.amountOfReloadsOnCurrentUrl, 0);
@@ -362,7 +361,7 @@ void main() {
   });
 
   testWidgets('evaluate Javascript', (WidgetTester tester) async {
-    WebViewController controller;
+    late WebViewController controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -375,15 +374,11 @@ void main() {
     expect(
         await controller.evaluateJavascript("fake js string"), "fake js string",
         reason: 'should get the argument');
-    expect(
-      () => controller.evaluateJavascript(null),
-      throwsA(anything),
-    );
   });
 
   testWidgets('evaluate Javascript with JavascriptMode disabled',
       (WidgetTester tester) async {
-    WebViewController controller;
+    late WebViewController controller;
     await tester.pumpWidget(
       WebView(
         initialUrl: 'https://flutter.io',
@@ -395,10 +390,6 @@ void main() {
     );
     expect(
       () => controller.evaluateJavascript('fake js string'),
-      throwsA(anything),
-    );
-    expect(
-      () => controller.evaluateJavascript(null),
       throwsA(anything),
     );
   });
@@ -444,7 +435,7 @@ void main() {
     );
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformWebView.javascriptChannelNames,
         unorderedEquals(<String>['Tts', 'Alarm']));
@@ -517,7 +508,7 @@ void main() {
     );
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformWebView.javascriptChannelNames,
         unorderedEquals(<String>['Tts', 'Alarm2', 'Alarm3']));
@@ -560,7 +551,7 @@ void main() {
     );
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformWebView.javascriptChannelNames,
         unorderedEquals(<String>['Tts']));
@@ -590,7 +581,7 @@ void main() {
     );
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(ttsMessagesReceived, isEmpty);
     expect(alarmMessagesReceived, isEmpty);
@@ -603,7 +594,7 @@ void main() {
 
   group('$PageStartedCallback', () {
     testWidgets('onPageStarted is not null', (WidgetTester tester) async {
-      String returnedUrl;
+      String? returnedUrl;
 
       await tester.pumpWidget(WebView(
         initialUrl: 'https://youtube.com',
@@ -613,7 +604,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       platformWebView.fakeOnPageStartedCallback();
 
@@ -627,7 +618,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       // The platform side will always invoke a call for onPageStarted. This is
       // to test that it does not crash on a null callback.
@@ -635,7 +626,7 @@ void main() {
     });
 
     testWidgets('onPageStarted changed', (WidgetTester tester) async {
-      String returnedUrl;
+      String? returnedUrl;
 
       await tester.pumpWidget(WebView(
         initialUrl: 'https://youtube.com',
@@ -650,7 +641,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       platformWebView.fakeOnPageStartedCallback();
 
@@ -660,7 +651,7 @@ void main() {
 
   group('$PageFinishedCallback', () {
     testWidgets('onPageFinished is not null', (WidgetTester tester) async {
-      String returnedUrl;
+      String? returnedUrl;
 
       await tester.pumpWidget(WebView(
         initialUrl: 'https://youtube.com',
@@ -670,7 +661,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       platformWebView.fakeOnPageFinishedCallback();
 
@@ -684,7 +675,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       // The platform side will always invoke a call for onPageFinished. This is
       // to test that it does not crash on a null callback.
@@ -692,7 +683,7 @@ void main() {
     });
 
     testWidgets('onPageFinished changed', (WidgetTester tester) async {
-      String returnedUrl;
+      String? returnedUrl;
 
       await tester.pumpWidget(WebView(
         initialUrl: 'https://youtube.com',
@@ -707,7 +698,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       platformWebView.fakeOnPageFinishedCallback();
 
@@ -722,13 +713,14 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       expect(platformWebView.hasNavigationDelegate, false);
 
       await tester.pumpWidget(WebView(
         initialUrl: 'https://youtube.com',
-        navigationDelegate: (NavigationRequest r) => null,
+        navigationDelegate: (NavigationRequest r) =>
+            NavigationDecision.navigate,
       ));
 
       expect(platformWebView.hasNavigationDelegate, true);
@@ -748,7 +740,7 @@ void main() {
           }));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       expect(platformWebView.hasNavigationDelegate, true);
 
@@ -773,7 +765,7 @@ void main() {
       ));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       expect(platformWebView.debuggingEnabled, true);
     });
@@ -782,7 +774,7 @@ void main() {
       await tester.pumpWidget(const WebView());
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       expect(platformWebView.debuggingEnabled, false);
     });
@@ -792,7 +784,7 @@ void main() {
       await tester.pumpWidget(WebView(key: key));
 
       final FakePlatformWebView platformWebView =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       await tester.pumpWidget(WebView(
         key: key,
@@ -826,8 +818,8 @@ void main() {
         ),
       );
 
-      final MyWebViewPlatform builder = WebView.platform;
-      final MyWebViewPlatformController platform = builder.lastPlatformBuilt;
+      final MyWebViewPlatform builder = WebView.platform as MyWebViewPlatform;
+      final MyWebViewPlatformController platform = builder.lastPlatformBuilt!;
 
       expect(
           platform.creationParams,
@@ -837,17 +829,14 @@ void main() {
               javascriptMode: JavascriptMode.disabled,
               hasNavigationDelegate: false,
               debuggingEnabled: false,
-              userAgent: WebSetting<String>.of(null),
+              userAgent: WebSetting<String?>.of(null),
               gestureNavigationEnabled: true,
             ),
-            // TODO(iskakaushik): Remove this when collection literals makes it to stable.
-            // ignore: prefer_collection_literals
-            javascriptChannelNames: Set<String>(),
           )));
     });
 
     testWidgets('loadUrl', (WidgetTester tester) async {
-      WebViewController controller;
+      late WebViewController controller;
       await tester.pumpWidget(
         WebView(
           initialUrl: 'https://youtube.com',
@@ -857,8 +846,8 @@ void main() {
         ),
       );
 
-      final MyWebViewPlatform builder = WebView.platform;
-      final MyWebViewPlatformController platform = builder.lastPlatformBuilt;
+      final MyWebViewPlatform builder = WebView.platform as MyWebViewPlatform;
+      final MyWebViewPlatformController platform = builder.lastPlatformBuilt!;
 
       final Map<String, String> headers = <String, String>{
         'header': 'value',
@@ -877,7 +866,7 @@ void main() {
     ));
 
     final FakePlatformWebView platformWebView =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformWebView.userAgent, isNull);
 
@@ -892,9 +881,9 @@ void main() {
 }
 
 class FakePlatformWebView {
-  FakePlatformWebView(int id, Map<dynamic, dynamic> params) {
+  FakePlatformWebView(int? id, Map<dynamic, dynamic> params) {
     if (params.containsKey('initialUrl')) {
-      final String initialUrl = params['initialUrl'];
+      final String? initialUrl = params['initialUrl'];
       if (initialUrl != null) {
         history.add(initialUrl);
         currentPosition++;
@@ -914,20 +903,20 @@ class FakePlatformWebView {
     channel.setMockMethodCallHandler(onMethodCall);
   }
 
-  MethodChannel channel;
+  late MethodChannel channel;
 
-  List<String> history = <String>[];
+  List<String?> history = <String?>[];
   int currentPosition = -1;
   int amountOfReloadsOnCurrentUrl = 0;
   bool hasCache = true;
 
-  String get currentUrl => history.isEmpty ? null : history[currentPosition];
-  JavascriptMode javascriptMode;
-  List<String> javascriptChannelNames;
+  String? get currentUrl => history.isEmpty ? null : history[currentPosition];
+  JavascriptMode? javascriptMode;
+  List<String>? javascriptChannelNames;
 
-  bool hasNavigationDelegate;
-  bool debuggingEnabled;
-  String userAgent;
+  bool? hasNavigationDelegate;
+  bool? debuggingEnabled;
+  String? userAgent;
 
   Future<dynamic> onMethodCall(MethodCall call) {
     switch (call.method) {
@@ -949,34 +938,28 @@ class FakePlatformWebView {
         break;
       case 'canGoBack':
         return Future<bool>.sync(() => currentPosition > 0);
-        break;
       case 'canGoForward':
         return Future<bool>.sync(() => currentPosition < history.length - 1);
-        break;
       case 'goBack':
         currentPosition = max(-1, currentPosition - 1);
         return Future<void>.sync(() {});
-        break;
       case 'goForward':
         currentPosition = min(history.length - 1, currentPosition + 1);
         return Future<void>.sync(() {});
       case 'reload':
         amountOfReloadsOnCurrentUrl++;
         return Future<void>.sync(() {});
-        break;
       case 'currentUrl':
         return Future<String>.value(currentUrl);
-        break;
       case 'evaluateJavascript':
         return Future<dynamic>.value(call.arguments);
-        break;
       case 'addJavascriptChannels':
         final List<String> channelNames = List<String>.from(call.arguments);
-        javascriptChannelNames.addAll(channelNames);
+        javascriptChannelNames!.addAll(channelNames);
         break;
       case 'removeJavascriptChannels':
         final List<String> channelNames = List<String>.from(call.arguments);
-        javascriptChannelNames
+        javascriptChannelNames!
             .removeWhere((String channel) => channelNames.contains(channel));
         break;
       case 'clearCache':
@@ -994,14 +977,14 @@ class FakePlatformWebView {
     };
     final ByteData data = codec
         .encodeMethodCall(MethodCall('javascriptChannelMessage', arguments));
-    ServicesBinding.instance.defaultBinaryMessenger
-        .handlePlatformMessage(channel.name, data, (ByteData data) {});
+    ServicesBinding.instance!.defaultBinaryMessenger
+        .handlePlatformMessage(channel.name, data, (ByteData? data) {});
   }
 
   // Fakes a main frame navigation that was initiated by the webview, e.g when
   // the user clicks a link in the currently loaded page.
   void fakeNavigate(String url) {
-    if (!hasNavigationDelegate) {
+    if (!hasNavigationDelegate!) {
       print('no navigation delegate');
       _loadUrl(url);
       return;
@@ -1013,9 +996,9 @@ class FakePlatformWebView {
     };
     final ByteData data =
         codec.encodeMethodCall(MethodCall('navigationRequest', arguments));
-    ServicesBinding.instance.defaultBinaryMessenger
-        .handlePlatformMessage(channel.name, data, (ByteData data) {
-      final bool allow = codec.decodeEnvelope(data);
+    ServicesBinding.instance!.defaultBinaryMessenger
+        .handlePlatformMessage(channel.name, data, (ByteData? data) {
+      final bool allow = codec.decodeEnvelope(data!);
       if (allow) {
         _loadUrl(url);
       }
@@ -1030,10 +1013,10 @@ class FakePlatformWebView {
       <dynamic, dynamic>{'url': currentUrl},
     ));
 
-    ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
+    ServicesBinding.instance!.defaultBinaryMessenger.handlePlatformMessage(
       channel.name,
       data,
-      (ByteData data) {},
+      (ByteData? data) {},
     );
   }
 
@@ -1045,14 +1028,14 @@ class FakePlatformWebView {
       <dynamic, dynamic>{'url': currentUrl},
     ));
 
-    ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
+    ServicesBinding.instance!.defaultBinaryMessenger.handlePlatformMessage(
       channel.name,
       data,
-      (ByteData data) {},
+      (ByteData? data) {},
     );
   }
 
-  void _loadUrl(String url) {
+  void _loadUrl(String? url) {
     history = history.sublist(0, currentPosition + 1);
     history.add(url);
     currentPosition++;
@@ -1061,13 +1044,13 @@ class FakePlatformWebView {
 }
 
 class _FakePlatformViewsController {
-  FakePlatformWebView lastCreatedView;
+  FakePlatformWebView? lastCreatedView;
 
   Future<dynamic> fakePlatformViewsMethodHandler(MethodCall call) {
     switch (call.method) {
       case 'create':
         final Map<dynamic, dynamic> args = call.arguments;
-        final Map<dynamic, dynamic> params = _decodeParams(args['params']);
+        final Map<dynamic, dynamic> params = _decodeParams(args['params'])!;
         lastCreatedView = FakePlatformWebView(
           args['id'],
           params,
@@ -1083,7 +1066,7 @@ class _FakePlatformViewsController {
   }
 }
 
-Map<dynamic, dynamic> _decodeParams(Uint8List paramsMessage) {
+Map<dynamic, dynamic>? _decodeParams(Uint8List paramsMessage) {
   final ByteBuffer buffer = paramsMessage.buffer;
   final ByteData messageBytes = buffer.asByteData(
     paramsMessage.offsetInBytes,
@@ -1114,9 +1097,8 @@ class _FakeCookieManager {
         return Future<bool>.sync(() {
           return hadCookies;
         });
-        break;
     }
-    return Future<bool>.sync(() => null);
+    return Future<bool>.sync(() => true);
   }
 
   void reset() {
@@ -1125,26 +1107,26 @@ class _FakeCookieManager {
 }
 
 class MyWebViewPlatform implements WebViewPlatform {
-  MyWebViewPlatformController lastPlatformBuilt;
+  MyWebViewPlatformController? lastPlatformBuilt;
 
   @override
   Widget build({
-    BuildContext context,
-    CreationParams creationParams,
-    @required WebViewPlatformCallbacksHandler webViewPlatformCallbacksHandler,
-    @required WebViewPlatformCreatedCallback onWebViewPlatformCreated,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
+    BuildContext? context,
+    CreationParams? creationParams,
+    required WebViewPlatformCallbacksHandler webViewPlatformCallbacksHandler,
+    WebViewPlatformCreatedCallback? onWebViewPlatformCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   }) {
     assert(onWebViewPlatformCreated != null);
     lastPlatformBuilt = MyWebViewPlatformController(
         creationParams, gestureRecognizers, webViewPlatformCallbacksHandler);
-    onWebViewPlatformCreated(lastPlatformBuilt);
+    onWebViewPlatformCreated!(lastPlatformBuilt);
     return Container();
   }
 
   @override
   Future<bool> clearCookies() {
-    return Future<bool>.sync(() => null);
+    return Future<bool>.sync(() => true);
   }
 }
 
@@ -1153,25 +1135,24 @@ class MyWebViewPlatformController extends WebViewPlatformController {
       WebViewPlatformCallbacksHandler platformHandler)
       : super(platformHandler);
 
-  CreationParams creationParams;
-  Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  CreationParams? creationParams;
+  Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers;
 
-  String lastUrlLoaded;
-  Map<String, String> lastRequestHeaders;
+  String? lastUrlLoaded;
+  Map<String, String>? lastRequestHeaders;
 
   @override
-  Future<void> loadUrl(String url, Map<String, String> headers) {
+  Future<void> loadUrl(String url, Map<String, String>? headers) async {
     equals(1, 1);
     lastUrlLoaded = url;
     lastRequestHeaders = headers;
-    return null;
   }
 }
 
 class MatchesWebSettings extends Matcher {
   MatchesWebSettings(this._webSettings);
 
-  final WebSettings _webSettings;
+  final WebSettings? _webSettings;
 
   @override
   Description describe(Description description) =>
@@ -1180,13 +1161,13 @@ class MatchesWebSettings extends Matcher {
   @override
   bool matches(
       covariant WebSettings webSettings, Map<dynamic, dynamic> matchState) {
-    return _webSettings.javascriptMode == webSettings.javascriptMode &&
-        _webSettings.hasNavigationDelegate ==
+    return _webSettings!.javascriptMode == webSettings.javascriptMode &&
+        _webSettings!.hasNavigationDelegate ==
             webSettings.hasNavigationDelegate &&
-        _webSettings.debuggingEnabled == webSettings.debuggingEnabled &&
-        _webSettings.gestureNavigationEnabled ==
+        _webSettings!.debuggingEnabled == webSettings.debuggingEnabled &&
+        _webSettings!.gestureNavigationEnabled ==
             webSettings.gestureNavigationEnabled &&
-        _webSettings.userAgent == webSettings.userAgent;
+        _webSettings!.userAgent == webSettings.userAgent;
   }
 }
 
@@ -1204,7 +1185,7 @@ class MatchesCreationParams extends Matcher {
       Map<dynamic, dynamic> matchState) {
     return _creationParams.initialUrl == creationParams.initialUrl &&
         MatchesWebSettings(_creationParams.webSettings)
-            .matches(creationParams.webSettings, matchState) &&
+            .matches(creationParams.webSettings!, matchState) &&
         orderedEquals(_creationParams.javascriptChannelNames)
             .matches(creationParams.javascriptChannelNames, matchState);
   }
